Conflicts:
	api/js/test.html
	resources/www_src/resources/commons/api/software.js

+++ main/src/main/java/com/bloatit/data/queries/DBRequests.java
@@ -47,7 +47,7 @@ public class DBRequests {
 
     /**
      * Make sure you test if the return is != null:
-     * 
+     *
      * <pre>
      * public static Team create() {
      *     DaoTeam dao = DBRequests.getById(DaoTeam.class, 12);
@@ -57,7 +57,7 @@ public class DBRequests {
      *     return new Team(dao);
      * }
      * </pre>
-     * 
+     *
      * @param <T>
      * @param persistant
      * @param id
@@ -79,7 +79,7 @@ public class DBRequests {
                 return null;
             }
         }
-        if (!persistant.isInstance(identifiable)){
+        if (!persistant.isInstance(identifiable)) {
             return null;
         }
         return identifiable;
@@ -98,12 +98,12 @@ public class DBRequests {
                                       SessionManager.createQuery("select count(*) from " + meta.getEntityName()));
     }
 
-    public static <T> int count(final Class<T> persistent) {
+    public static <T> Long count(final Class<T> persistent) {
         final ClassMetadata meta = SessionManager.getSessionFactory().getClassMetadata(persistent);
         return ((Long) SessionManager.getSessionFactory()
                                      .getCurrentSession()
                                      .createQuery("select count(*) from " + meta.getEntityName())
-                                     .uniqueResult()).intValue();
+                                     .uniqueResult());
     }
 
     public static PageIterable<DaoFeature> featuresOrderByPopularity() {
